Better URL handling
New environment variable JMAP_HOSTNAME which defaults to
`betajmap.fastmail.com` for now.

auth_uri is then calculated from that, using the proper autodiscovery
mechanism for the rest of the URL, as specified for JMAP.

api_url is also unhardcoded, in favour of the apiUrl key from the
session object.