get python3 hello world working
This is an update of the hello-world.py script that works. the previous
version had an issue with identityId being an invalidProperty. In fact,
as @rjbs pointed out, it also has to do with the envelope. So this
version also defines an envelope along with an identityId (which also
needs to be set).